Implement bug fix to reset button
Previously, it was possible to click the reset button multiple times before the .showPattern
method was called for the new game. This made it visually bug out, so that was no good.

The fix wasn't as hard as I expected. The trick was to figure out the correct timeout delay for
making the jQBoardNode unclickable in both the startNewGame and showPattern methods.